slidewindow data and graphs
1. There are two parts of this code:
a. To get data 'alignsession1' and 'alignsession2'. From line 1 to line 135 is getting spiking rate of each window of each electrode, for session 1 and session 2 respectively. Only for original data, not for standardized data. From line 136 to line 157 is to get word type of each trial.
b. From line 158 to line 177 are doing slide window one way ANOVA and sort the result in 'pslide1','pslide2','Fslide1' and 'Fslide2'.
c. draw slide window p-values of electrodes which has shown significance in global one-way ANOVA in file 'onewaydata'.

2. How to use this file?
a. This file should be used after running 'onewaydata'. 'onewaydata' has 5 parts as I wrote in the description of 'onewaydata', and part a is time consuming, say 4-5 min. You can save the result of part a, and only run part b-e each time you want to use 'slidewindow', and this is pretty fast, less than half min.
b. need to enter the file name you want to load in Line 5, 39, 71 and 105. 
b. In Line 180, you need to set i = 1:??, the number ?? comes from length(sigslct1org). sigslct1org can be obtained by running 'onewaydata'. Then in Line 181, you need to set the subplot numbers. Line 180 and 181 are for session 1, and you need to do the samething to Line 192 (the ?? should come from length(sigslct2org)) and Line 193.

3. The result you get.
a. 'alignsession1' includes info such as spiking rates of each window and wordtype for all 34 words of each electrode in session 1, and 'alignsession2' for session2.
b. 'pslide1' is storing the sliding p values of electrodes which has shown significance in global ANOVA when running 'onewaydata' of original data. 'pslide2' is same as above but for session 2.
c. 'slide_window_p_session1.gif' is the plot of 'pslide1', and 'slide_windwow_p_session2.gif' is the plot of 'pslide2'.